aob-30: Change from Integer ids to Long ids.

+++ src/main/java/project11/amazinbookstore/user/UserRepository.java
@@ -7,6 +7,6 @@ import project11.amazinbookstore.user.RegisteredUser;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<RegisteredUser, Integer> {
+public interface UserRepository extends JpaRepository<RegisteredUser, Long> {
     Optional<RegisteredUser> findByUsername(String username);
 }
